requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0
django-debug-toolbar==4.2.0
django-redis==5.4.0
djangorestframework==3.14.0
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytz==2024.1
redis==5.0.1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
timedelta-isoformat==0.6.2.11
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
whitenoise==6.6.0
